+++ source/dsp/GranularEngine.h
@@ -1,13 +1,3 @@
-/*
-  ==============================================================================
-
-	GranularEngine.h
-	Created: 15 Feb 2023 1:44:09pm
-	Author:  user
-
-  ==============================================================================
-*/
-
 // http://www.rossbencina.com/static/code/granular-synthesis/BencinaAudioAnecdotes310801.pdf
 
 // Top level container that orchestrate all the blocks
@@ -17,47 +7,42 @@
 #include <juce_audio_basics/juce_audio_basics.h>
 #include <juce_dsp/juce_dsp.h>
 
-#include "../framework/PluginParams.h"
-#include "../framework/PluginTypes.h"
-#include "../framework/bridge/LockFreeDoubleBuffer.h"
-#include "../utils/AtomicSharedPtr.h"
-#include "../utils/struct/SmoothedParameters.h"
-#include "../utils/struct/VisualSnapshot.h"
 #include "GrainEnvelope.h"
 #include "GrainPool.h"
 #include "GrainProcessor.h"
 #include "PositionModulator.h"
 #include "Scheduler.h"
+#include "framework/PluginParams.h"
+#include "framework/PluginTypes.h"
+#include "framework/bridge/PingPongBuffer.h"
+#include "utils/AtomicSharedPtr.h"
+#include "utils/struct/SmoothedParameters.h"
+#include "utils/struct/VisualSnapshot.h"
 
 namespace particules
 {
-    //class VisualSnapshot;
     struct ParameterSnapshot;
     class AudioState;
     class GranularEngine
     {
     public:
-        GranularEngine(LockFreeDoubleBuffer<VisualSnapshot>& vb, AudioState& as);
+        GranularEngine(PingPongBuffer<VisualSnapshot>& vb, AudioState& as);
         ~GranularEngine() = default;
 
-        void process(
-            AudioBuffer& output, int bufferSize, float* const* outputPtrs, int outputNumChannels, const ParameterSnapshot& ps);
+        void process(AudioBuffer& outputBuffer, AudioBuffer& inputBuffer, int bufferSize, float* const* outputPtrs,
+            int outputNumChannels, const ParameterSnapshot& ps);
+
         void init(double, int, int);
+        void clear() noexcept;
         int getNumActiveGrains() const noexcept { return pool.getNumActiveGrains(); }
 
-        void setInputBuffer(std::shared_ptr<const AudioBuffer> ptr) noexcept { inputBufferPtr.store(std::move(ptr)); }
-        std::shared_ptr<const AudioBuffer> getInputBuffer() const noexcept { return inputBufferPtr.load(); }
-        const bool isInputBufferLoaded() const noexcept { return inputBufferPtr.load() != nullptr; }
-
     private:
-        void writeVisualSnapshot() noexcept ;
+        void writeVisualSnapshot() noexcept;
         void gainProcess(juce::dsp::ProcessContextReplacing<float>, const float);
         void updateSmoothedParameters() noexcept;
         void setTargetSmoothedValue(const ParameterSnapshot&) noexcept;
 
         static constexpr uint8_t mMaxEvent = params::maxSpawnsPerBlock;
-
-        AtomicSharedPtr<const AudioBuffer> inputBufferPtr; // should be downmixed
 
         // core data members
         const float refreshRate;
@@ -83,8 +68,7 @@ namespace particules
         SmoothedParameters smoothedParams;
 
         AudioState& audioState;
-        LockFreeDoubleBuffer<VisualSnapshot>& visualBuffer;
-
+        PingPongBuffer<VisualSnapshot>& visualBuffer;
     };
 }
 //juce::dsp::DryWetMixer<float> mixerProcessor;