+++ source/framework/PluginTypes.h
@@ -14,6 +14,6 @@ namespace particules
     using ValueTreeState = juce::AudioProcessorValueTreeState; // juce::juce_audio_processors
     using AudioBlock = juce::dsp::AudioBlock<float>; // juce::juce_dsp
     using AudioBuffer = juce::AudioBuffer<float>; // juce::juce_audio_basics
-    using AudioLoadedCallback =std::function<void(AudioBuffer&, const juce::File&)>; // juce::juce_audio_basics
+    using AudioLoadedCallback =std::function<void(std::unique_ptr<AudioBuffer>, const juce::File&)>; // juce::juce_audio_basics
     using str = juce::String; // juce::juce_core
 }