+++ source/framework/bridge/RingBuffer.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "../Core.h"
 
-// SPSC Queue (Ring Buffer)
+// SPSC Queue (FIFO)
 // its move pointers between thread
 // gonna be used for SPSC pattern
 // specially for moving AudioBuffer
@@ -10,10 +10,10 @@
 namespace particules
 {
     template <typename T, size_t Size = 16>
-    class LockFreePointerQueue
+    class RingBuffer
     {
     public:
-        LockFreePointerQueue() = default;
+        RingBuffer() = default;
 
         bool push(T* item) noexcept
         {
@@ -21,7 +21,7 @@ namespace particules
             const int nextWrite = (currentWrite + 1) % Size;
 
             if(nextWrite == readPos.load(std::memory_order_acquire))
-                return false; // File pleine
+                return false; // queue is full
 
             buffer[currentWrite] = item;
             writePos.store(nextWrite, std::memory_order_release);
@@ -33,7 +33,7 @@ namespace particules
             const int currentRead = readPos.load(std::memory_order_relaxed);
 
             if(currentRead == writePos.load(std::memory_order_acquire))
-                return nullptr; // File vide
+                return nullptr; // queue is empty
 
             T* item = buffer[currentRead];
             readPos.store((currentRead + 1) % Size, std::memory_order_release);