+++ source/gui_app/GuiAppProcessor.h
@@ -4,15 +4,15 @@
 #include <juce_audio_processors/juce_audio_processors.h> // audio processor
 #include <juce_core/juce_core.h> // memory block and string
 
-#include "framework/bridge/LockFreeDoubleBuffer.h"
 #include "../framework/PluginTypes.h"
 #include "../framework/bridge/AudioState.h"
+#include "framework/bridge/PingPongBuffer.h"
 //#include "../framework/bridge/GrainVisualBuffer.h"
 #include "../framework/bridge/ParameterView.h"
 #include "../framework/bridge/UIState.h"
 #include "../utils/struct/ProcessorFacade.h"
-#include "../utils/struct/VisualSnapshot.h"
 #include "../utils/struct/UIContext.h"
+#include "../utils/struct/VisualSnapshot.h"
 
 // just to test GUI separated than the audio DSP
 namespace particules
@@ -31,8 +31,8 @@ namespace particules
                 .withValueFromStringFunction(valueFromStringFunc));
     }
 
-    inline std::unique_ptr<juce::AudioParameterFloat> createNormalizedParameter(const juce::ParameterID& id, const str& name,
-        float min, float max, float skew = 0.5f, float init = 0.5f)
+    inline std::unique_ptr<juce::AudioParameterFloat> createNormalizedParameter(
+        const juce::ParameterID& id, const str& name, float min, float max, float skew = 0.5f, float init = 0.5f)
     {
         juce::NormalisableRange<float> range{min, max};
         range.setSkewForCentre(skew);
@@ -74,12 +74,11 @@ namespace particules
         UIContext& getUIContext() noexcept { return uic; };
 
     private:
-        LockFreeDoubleBuffer<VisualSnapshot> visualBuffer;
+        PingPongBuffer<VisualSnapshot> visualBuffer;
         ValueTreeState apvts;
         AudioState audioState;
         ParameterView paramsView;
         UIState uiState;
-
         ProcessorFacade facade;
         UIContext uic;