+++ source/utils/struct/AudioPayload.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <memory>
+
+#include "framework/PluginTypes.h"
+
+namespace particules
+{
+    struct AudioPayload
+    {
+        std::unique_ptr<AudioBuffer> buffer;
+        int numSamples = 0;
+        int numChannels = 0;
+        juce::File file;
+    };
+}